requirements.txt
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.12.2
django-import-export==2.5.0
pytz==2019.3
sqlparse==0.2.4
